Add support for watchify
Previously Jadeify only registered the required Jade file, not allowing watchify to rebuild or cache changes to the Jade dependency tree.  This modification uses the Jade parser to create an array of the dependency tree of a required Jade file and registers its dependencies as "file" events for browserify and watchify.